Add Eloquent Sluggable package for automatic slug generation
Installed the "Cviebrock\EloquentSluggable" package to automatically generate slugs from the title field in the Post model. Published the package configuration file and updated the Post model to use the Sluggable trait for easy slug creation. The slug will now be automatically generated when a new post is created or the title is updated.